[hotfix][tests] Extract SubmittedJobGraphStore implementation from JobManagerHARecoveryTest
